test(sgwallet): refactor wallet test to make write test case easier.
