+++ sgwallet/tests/wallet_test_by_rpc_chain.rs
@@ -1,36 +1,37 @@
 #[macro_use]
 extern crate rusty_fork;
-use crate::wallet_test_helper::{test_deploy_custom_module, test_wallet};
+use crate::wallet_test_helper::{
+    test_channel_event_watcher_async, test_deploy_custom_module, test_wallet_async,
+};
 use common::with_wallet;
-use failure::prelude::*;
 use libra_logger::prelude::*;
 use sgchain::star_chain_client::{ChainClient, StarChainClient};
 use std::sync::Arc;
-mod common;
+pub mod common;
 mod transfer;
-mod wallet_test_helper;
+pub mod wallet_test_helper;
 
 rusty_fork_test! {
 
     #[test]
     fn test_wallet_with_rpc_client() {
-        run_test_wallet_with_rpc_client().unwrap();
+        run_test_wallet_with_rpc_client();
     }
 
     #[test]
     fn test_deploy_custom_module_by_rpc_client() {
-        run_deploy_custom_module_by_rpc_client().unwrap();
+        run_deploy_custom_module_by_rpc_client();
+    }
+
+    #[test]
+    fn test_wallet_transfer_htlc() {
+        run_test_wallet_transfer_htlc();
     }
 }
 
 #[test]
-// just for manual execute
-fn test_wallet_with_rpc_client_manual() {
-    run_test_wallet_with_rpc_client().unwrap();
-}
-
-#[test]
-fn test_wallet_transfer_htlc() {
+#[ignore]
+fn run_test_wallet_transfer_htlc() {
     let result = run_with_rpc_client(|chain_client| {
         with_wallet(chain_client, |rt, sender, receiver| {
             rt.block_on(transfer::transfer_htlc(sender.clone(), receiver.clone()))
@@ -42,23 +43,39 @@ fn test_wallet_transfer_htlc() {
     }
 }
 
-fn run_test_wallet_with_rpc_client() -> Result<()> {
-    let (_config, _logger, _handler) = sgchain::main_node::run_node(None, false);
-    info!("note is running.");
-    //TODO use a random port and check port available for port conflict.
-    let rpc_client = StarChainClient::new("127.0.0.1", 8000);
-    let chain_client = Arc::new(rpc_client);
-    test_wallet(chain_client)?;
-    Ok(())
+#[test]
+#[ignore]
+fn run_test_wallet_with_rpc_client() {
+    if let Err(e) = run_with_rpc_client(|chain_client| {
+        with_wallet(chain_client, |rt, sender, receiver| {
+            rt.block_on(test_wallet_async(sender.clone(), receiver.clone()))
+        })
+    }) {
+        error!("err: {:#?}", e);
+        assert!(false);
+    }
 }
 
-fn run_deploy_custom_module_by_rpc_client() -> Result<()> {
-    let (_config, _logger, _handler) = sgchain::main_node::run_node(None, false);
-    info!("note is running.");
-    let rpc_client = StarChainClient::new("127.0.0.1", 8000);
-    let chain_client = Arc::new(rpc_client);
-    test_deploy_custom_module(chain_client)?;
-    Ok(())
+#[test]
+#[ignore]
+fn run_deploy_custom_module_by_rpc_client() {
+    if let Err(e) = run_with_rpc_client(|chain_client| test_deploy_custom_module(chain_client)) {
+        error!("err: {:#?}", e);
+        assert!(false);
+    }
+}
+
+#[test]
+#[ignore]
+fn run_test_channel_event_watcher() {
+    if let Err(e) = run_with_rpc_client(|chain_client| {
+        common::with_wallet(chain_client, |rt, sender, receiver| {
+            rt.block_on(test_channel_event_watcher_async(sender, receiver))
+        })
+    }) {
+        error!("err: {:#?}", e);
+        assert!(false);
+    }
 }
 
 fn run_with_rpc_client<F, T>(mut f: F) -> T
@@ -66,7 +83,7 @@ where
     F: FnMut(Arc<dyn ChainClient>) -> T,
 {
     let (_config, _logger, _handler) = sgchain::main_node::run_node(None, false);
-    info!("note is running.");
+    info!("node is running.");
     let rpc_client = StarChainClient::new("127.0.0.1", 8000);
     f(Arc::new(rpc_client))
 }
